Organize drawing
