+++ draw/world.h
@@ -1,0 +1,4 @@
+#pragma once
+#include <draw/draw.h>
+
+void draw_world_example(Canvas *canvas);
